Only reset grace period size on renew when expiration is further in the future than before
